+++ src/UDPsView.js
@@ -1,6 +1,5 @@
 import { useState } from 'react';
 import PropTypes from 'prop-types';
-import { noop } from 'lodash';
 import { FormattedMessage } from 'react-intl';
 import {
   MultiColumnList,
@@ -25,7 +24,6 @@ import css from './UDPSearch.css';
 
 const UDPsView = ({
   children,
-  contentRef,
   data = {},
   onNeedMoreData,
   onSelectRow,
@@ -86,7 +84,7 @@ const UDPsView = ({
           source={source}
           searchTerm={query.query || ''}
           filterPaneIsVisible
-          toggleFilterPane={noop}
+          toggleFilterPane={toggleFilterPane}
         />
       </div>
     );
@@ -123,7 +121,7 @@ const UDPsView = ({
   };
 
   return (
-    <div data-test-udp-instances ref={contentRef}>
+    <div>
       <SearchAndSortQuery
         initialFilterState={{}}
         initialSearchState={{ query: '' }}
@@ -171,7 +169,6 @@ const UDPsView = ({
                             aria-label={ariaLabel}
                             autoFocus
                             className={css.searchField}
-                            data-test-udp-search-input
                             id="input-udp-search"
                             inputRef={searchField}
                             name="query"
@@ -195,7 +192,6 @@ const UDPsView = ({
                         fullWidth
                         id="clickable-search-udps"
                         type="submit"
-                        data-test-udp-search-submit
                       >
                         <FormattedMessage id="stripes-smart-components.search" />
                       </Button>
@@ -259,7 +255,6 @@ const UDPsView = ({
 
 UDPsView.propTypes = {
   children: PropTypes.object,
-  contentRef: PropTypes.object,
   data: PropTypes.object,
   onNeedMoreData: PropTypes.func,
   onSelectRow: PropTypes.func,