+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix GitHub Actions workflow not running for tags ([FOLIO-4086](https://folio-org.atlassian.net/browse/FOLIO-4086))
 * Add support for interface `aggregator-settings 2.0` ([UIPFUDP-42](https://folio-org.atlassian.net/browse/UIPFUDP-42))
 * Add support for interface `erm-usage-harvester 2.0`([UIPFUDP-43](https://folio-org.atlassian.net/browse/UIPFUDP-43))
+* Deleting search string via keyboard does not update search results ([UIPFUDP-36](https://folio-org.atlassian.net/browse/UIPFUDP-36))
 
 ## 6.1.0 (2024-03-20)
 * Translation updates