+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change history for ui-plugin-find-erm-usage-data-provider
 
-## 7.0.0 (IN PROGRESS)
+## 7.1.0 (IN PROGRESS)
+
+## [7.0.0](https://github.com/folio-org/ui-plugin-find-erm-usage-data-provider/tree/v7.0.0) (2025-03-12)
 * Remove unused code ([UIPFUDP-39](https://folio-org.atlassian.net/browse/UIPFUDP-39))
 * Migrate react-intl to v7 ([UIPFUDP-46](https://folio-org.atlassian.net/browse/UIPFUDP-46))
 * Migrate stripes dependencies to their Sunflower versions ([UIPFUDP-45](https://folio-org.atlassian.net/browse/UIPFUDP-45))